requirements.txt
numpy
pandas
scikit-learn
opencv-python
Pillow
tensorflow
streamlit
joblib
textblob
matplotlib
tqdm
